Module Odoc_odoc.Compile

type package_spec = {

package : string;

output : Fpath.t;

}

type parent_spec = {

parent : string option;

children : string list;

output : Fpath.t;

}

type parent_id_spec = {

parent_id : string;

output_dir : string;

}

type cli_spec =

| CliNoParent of Fpath.t

| CliPackage of package_spec

| CliParent of parent_spec

| CliParentId of parent_id_spec

val name_of_output : prefix:string -> Fs.File.t -> string

Compute the name of the page from the output file. Prefix is the prefix to remove from the filename.

val resolve_imports : Resolver.t -> Odoc_model.Lang.Compilation_unit.Import.t list -> Odoc_model.Lang.Compilation_unit.Import.t list

val resolve_parent_page : Resolver.t -> string -> (Odoc_model.Paths.Identifier.ContainerPage.t * Odoc_model.Lang.Page.child list, [> Or_error.msg ]) Or_error.result

Parse and resolve a parent reference. Returns the identifier of the parent and its children as a list of reference.

val mk_id : string -> Odoc_model.Paths.Identifier.ContainerPage.t option

val path_of_id : string -> Odoc_model.Paths.Identifier.ContainerPage.t option -> Fpath.t

val compile : resolver:Resolver.t -> hidden:bool -> cli_spec:cli_spec -> warnings_options:Odoc_model.Error.warnings_options -> short_title:string option -> Fpath.t -> (unit, [> Or_error.msg ]) Or_error.result

Produces .odoc files out of .cm{i,t,ti} or .mld files.