Module Fs.Directory

type t = directory

val dirname : t -> t

val basename : t -> t

val append : t -> t -> t

val reach_from : dir:t -> string -> t

raises Invalid_argument
if parent/name exists but is not a directory.

val contains : parentdir:t -> file -> bool

val compare : t -> t -> int

val mkdir_p : t -> unit

val of_file : file -> t

val of_string : string -> t

val to_string : t -> string

val to_fpath : t -> Fpath.t

val fold_files_rec : ?ext:string -> ('a -> file -> 'a) -> 'a -> t -> 'a

fold_files_rec_result ~ext f acc d recursively folds f over the files with extension matching ext (defaults to "") contained in d and its sub directories.

val fold_files_rec_result : ?ext:string -> ('a -> file -> ('a, Or_error.msg) Or_error.result) -> 'a -> t -> ('a, [> Or_error.msg ]) Or_error.result

fold_files_rec_result ~ext f acc d recursively folds f over the files with extension matching ext (defaults to "") contained in d and its sub directories. Stop as soon as f returns Error _.

module Table : Hashtbl.S with type key = t