Module Odoc_odoc.Odoc_file

Load and save .odoc and .odocl files.

type content =

| Page_content of Odoc_model.Lang.Page.t

| Impl_content of Odoc_model.Lang.Implementation.t

| Unit_content of Odoc_model.Lang.Compilation_unit.t

| Asset_content of Odoc_model.Lang.Asset.t

Either a page or a module or something else.

type t = {

content : content;

warnings : Odoc_model.Error.t list;

}

Serialization

val save_page : Fs.File.t -> warnings:Odoc_model.Error.t list -> Odoc_model.Lang.Page.t -> unit

Save a page. The page- prefix is added to the file name if missing.

val save_unit : Fs.File.t -> warnings:Odoc_model.Error.t list -> Odoc_model.Lang.Compilation_unit.t -> unit

Save a module.

val save_impl : Fs.File.t -> warnings:Odoc_model.Error.t list -> Odoc_model.Lang.Implementation.t -> unit

Save an implementation. The src- prefix is added to the file name if missing.

Deserialization

val load : Fs.File.t -> (t, [> Or_error.msg ]) Or_error.result

Load an .odoc file.

val load_root : Fs.File.t -> (Odoc_model.Root.t, [> Or_error.msg ]) Or_error.result

Only load the root. Faster than load, used for looking up imports.

val save_index : Fs.File.t -> Odoc_index.t -> unit

val load_index : Fs.File.t -> (Odoc_index.t, [> Or_error.msg ]) Or_error.result

Load a .odoc-index file.

val save_sidebar : Fs.File.t -> Odoc_document.Sidebar.t -> unit

val load_sidebar : Fs.File.t -> (Odoc_document.Sidebar.t, [> Or_error.msg ]) Or_error.result

Load a .odoc-index file.

val save_asset : Fpath.t -> warnings:Odoc_model.Error.t list -> Odoc_model.Lang.Asset.t -> unit